requirements-example.txt
fastapi
uvicorn
pyodbc
redis
pydantic
pyyaml